% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-model-matrix.R
\name{compute_model_matrix}
\alias{compute_model_matrix}
\title{Compute model matrix}
\usage{
compute_model_matrix(parent_dir, max_sample = Inf, sample_seed = 1,
  poly_degree = 4L, na_val = 0, filter_corpus = NULL, ltm = TRUE,
  viewpoint_dir = file.path(parent_dir, "0-viewpoints"),
  ppm_dir = file.path(parent_dir, "1-ppm"),
  output_dir = file.path(parent_dir, "2-model-matrix"),
  viewpoints = read_viewpoints(viewpoint_dir),
  seq_test = list_seq_test(ppm_dir), allow_repeats = FALSE)
}
\arguments{
\item{parent_dir}{(Character scalar)
The parent directory for the output files, shared with functions such as
\code{\link{compute_viewpoints}} and \code{\link{compute_ppm_analyses}}.
Ignored if all other directory arguments are manually specified.}

\item{max_sample}{(Numeric scalar)
Maximum number of events to sample for the model matrix,
defaults to \code{Inf} (no downsampling).
Lower values of \code{max_sample} prompt random downsampling.}

\item{sample_seed}{(Integer scalar)
Random seed to make the downsampling reproducible.}

\item{poly_degree}{(Integer scalar)
Degree of the polynomials to compute for the continuous features.}

\item{na_val}{(Numeric scalar)
Value to use to code for NA in the model matrix.
The statistical analyses are mostly unaffected by this value.}

\item{filter_corpus}{(NULL or a function)
An optional function to apply to the corpus to determine which
events should be retained in the model matrix.
The function is applied to the corpus object saved as \code{corpus.rds}
in \code{viewpoint_dir}. This corpus object takes the form of a
\code{\link[tibble]{tibble}}; the function should return a
row-subset of this \code{tibble}.}

\item{ltm}{(Logical scalar, default = \code{TRUE})
If \code{FALSE}, long-term (i.e. pretrained) PPM model outputs are
excluded from the model matrix.}

\item{viewpoint_dir}{(Character scalar)
The directory for the already-generated
output files from \code{\link{compute_viewpoints}}.
The default should be correct if the user used the
default \code{dir} argument in \code{\link{compute_viewpoints}}.}

\item{output_dir}{(Character scalar)
The output directory for the model matrix.
Will be created if it doesn't exist already.}

\item{viewpoints}{Character vector listing the viewpoints to be included in the model matrix.
By default this list is read from \code{viewpoint_dir}.}

\item{seq_test}{Integer vector identifying which sequences should be sampled from for
constructing the model matrix, indexing into the \code{corpus}
argument of \code{\link{compute_viewpoints}}.
Defaults to the \code{seq_test} argument that was provided to
\code{\link{compute_viewpoints}}.}

\item{allow_repeats}{(Logical scalar)
Whether repeated chords are theoretically permitted in the
chord sequences. It is recommended to remove such repetitions
before modelling.}

\item{compute_ppm_analyses}{(Character scalar)
The directory for the already-generated
output files from \code{\link{compute_ppm_analyses}}.
The default should be correct if the user used the
default \code{dir} argument in \code{\link{compute_ppm_analyses}}.}
}
\value{
The primary output is written to disk in the \code{dir} directory.
The model matrix provides metafeature values
(i.e. expectedness values for discrete features
and polynomial values for continuous features)
over the entire chord alphabet at every location in \code{seq_test}.
}
\description{
Computes the model matrix, which compiles together expectedness values
from the PPM analyses as well as polynomial expansions of the
continuous features.
}
\details{
The following routines should have been run already:
\enumerate{
\item \code{\link{compute_viewpoints}}
\item \code{\link{compute_ppm_analyses}}
}
}
