% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-compute-ppm-analyses.R
\name{compute_ppm_analyses}
\alias{compute_ppm_analyses}
\title{Compute PPM analyses}
\usage{
compute_ppm_analyses(parent_dir, viewpoint_dir = file.path(parent_dir,
  "0-viewpoints"), output_dir = file.path(parent_dir, "1-ppm"),
  stm_opt = stm_options(), ltm_opt = ltm_options(),
  seq_test_folds = list(readRDS(file.path(viewpoint_dir,
  "about.rds"))$seq_test),
  seq_pretrain = readRDS(file.path(viewpoint_dir,
  "about.rds"))$seq_pretrain,
  viewpoints = readRDS(file.path(viewpoint_dir,
  "about.rds"))$discrete_viewpoints)
}
\arguments{
\item{parent_dir}{(Character scalar)
The parent directory for the output files, shared with functions such as
\code{\link{compute_viewpoints}} and \code{\link{compute_model_matrix}}.
Ignored if all other directory arguments are manually specified.}

\item{viewpoint_dir}{(Character scalar)
The directory for the already-generated
output files from \code{\link{compute_viewpoints}}.
The default should be correct if the user used the
default \code{dir} argument in \code{\link{compute_viewpoints}}.}

\item{output_dir}{(Character scalar)
The output directory for the PPM analyses.
Will be created if it doesn't exist already.}

\item{stm_opt}{Options list for the short-term PPM models, as created by the function
\code{\link{stm_options}}.}

\item{ltm_opt}{Options list for the long-term PPM models, as created by the function
\code{\link{ltm_options}}.}

\item{seq_test_folds}{List of cross-validation folds for the test sequences.
Each fold is represented as an integer vector,
with the integers indexing the sequences within the corpus
(see \code{\link{compute_viewpoints}}).
The algorithm iterates over each fold, predicting the sequences
within that fold, and training the model using the combination of
a) the sequences from the other folds in \code{seq_test_folds} and
b) the sequences identified in \code{seq_pretrain}.
By default, there is just one fold corresponding to the \code{seq_test}
argument of \code{\link{compute_viewpoints}}.}

\item{seq_pretrain}{(Integer vector)
Sequences used to pretrain the model (in addition to any cross-validation training
specified by \code{seq_test_folds}), specified as integer indices
of the corpus.
Defaults to the complement of \code{seq_test} as specified in
\code{\link{compute_viewpoints}}.}

\item{viewpoints}{List of discrete viewpoints to analyse,
in the format produced by the \code{$discrete_viewpoints} slot
of the \code{about.rds} file produced by \code{\link{compute_viewpoints}}.
Defaults to the full set of discrete viewpoints as specified
in \code{\link{compute_viewpoints}}.}
}
\value{
The primary output is written to disk in the \code{dir} directory.
The output matrices provide raw probabilities for each event in the
chord alphabet.
}
\description{
This function models discrete viewpoints using the
Prediction by Partial Match (PPM) algorithm.
}
\details{
\code{\link{compute_viewpoints}} should be run first.
By default, only sequences in
\code{seq_test} (see \code{\link{compute_viewpoints}})
are modelled using PPM.
The default PPM implementation corresponds to that described in
\insertCite{Pearce2005;textual}{hvr}.
}
