% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-compute-predictions.R
\name{compute_predictions}
\alias{compute_predictions}
\title{Compute predictions}
\usage{
compute_predictions(
  parent_dir,
  weights = NULL,
  model_matrix_dir = file.path(parent_dir, "2-model-matrix"),
  regression_dir = file.path(parent_dir, "3-viewpoint-regression"),
  output_dir = file.path(parent_dir, "4-predictions")
)
}
\arguments{
\item{parent_dir}{(Character scalar)
The parent directory for the output files, shared with functions such as
\code{\link{compute_viewpoints}} and \code{\link{compute_ppm_analyses}}.
Ignored if all other directory arguments are manually specified.}

\item{weights}{If \code{NULL} (default),
the regression weights are extracted from the viewpoint regression model
output in \code{regression_dir}.
Alternatively, the weights can be manually specified as a named numeric vector
where the names correspond to the \code{label} column of the
\code{\link[tibble]{tibble}} stored in \code{predictors.rds}
in \code{model_matrix_dir}.}

\item{model_matrix_dir}{(Character scalar)
The directory for the already-generated
output files from \code{\link{compute_model_matrix}}.
The default should be correct if the user used the
default \code{dir} argument in \code{\link{compute_model_matrix}}.}

\item{regression_dir}{(Character scalar)
The directory for the already-generated
output files from \code{\link{viewpoint_regression}}.
The default should be correct if the user used the
default \code{dir} argument in \code{\link{viewpoint_regression}}.}

\item{output_dir}{(Character scalar)
The output directory for the new results.
Will be created if it doesn't exist already.}
}
\value{
A \code{\link[tibble]{tibble}} with one row for each event in the
model matrix, with the following columns:
\itemize{
\item \code{seq_id} - Index of the sequence within the corpus
(see \code{\link{compute_viewpoints}}).
\item \code{event_id} - Index of the chord within the sequence.
\item \code{chord_id} - Encoded chord symbol
(see \code{\link[hrep]{pc_chord}}, \code{\link[hrep]{encode}}).
\item \code{chord} - Decoded chord symbol.
\item \code{probability} - Probability of the observed chord.
\item \code{information_content} - Information content of the observed chord.
}
}
\description{
Computes predictions from a trained viewpoint regression model.
The following routines will typically be run first:
\enumerate{
\item \code{\link{compute_viewpoints}}
\item \code{\link{compute_ppm_analyses}}
\item \code{\link{compute_model_matrix}}
\item \code{\link{viewpoint_regression}}
}
}
