% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-compute-viewpoints.R
\name{compute_viewpoints}
\alias{compute_viewpoints}
\title{Compute viewpoints}
\usage{
compute_viewpoints(corpus, parent_dir, output_dir = file.path(parent_dir,
  "0-viewpoints"), seq_test = seq_along(corpus),
  viewpoints = hvr::hvr_viewpoints, verbose = TRUE)
}
\arguments{
\item{corpus}{A corpus of chord sequences as created by \code{\link[hrep]{corpus}}.
Must be encoded in the \code{\link[hrep]{pc_chord}} representation.}

\item{parent_dir}{(Character scalar)
The parent directory for the output files, shared with functions such as
\code{\link{compute_ppm_analyses}} and \code{\link{compute_model_matrix}}.
Ignored if \code{dir} is provided.}

\item{output_dir}{(Character scalar)
The output directory for the viewpoint analyses.
Will be created if it doesn't exist already.}

\item{seq_test}{(Integer vector)
Provides the indices of the sequences in \code{corpus} that will
eventually be predicted by the model.
Sequences not included in \code{seq_test} will only be used for training.
Training sequences only need viewpoints to be computed for observed chords,
not the full alphabet of possible chords.}

\item{viewpoints}{List of viewpoints to apply, as created by \code{\link{new_viewpoint}}.
Defaults to a fairly comprehensive list, \code{\link{hvr_viewpoints}}.}

\item{verbose}{(Logical scalar)
Whether or not to print verbose status reports.}
}
\value{
The primary output is written to disk in the \code{output_dir} directory.
}
\description{
This function computes viewpoints (features) for every chord sequence
in a corpus, and saves the result to disk.
}
