% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-viewpoint-regression.R
\name{get_costs}
\alias{get_costs}
\alias{get_costs.viewpoint_regression}
\title{Get costs}
\usage{
get_costs(x)

\method{get_costs}{viewpoint_regression}(x)
}
\arguments{
\item{x}{Viewpoint regression model as created with
\code{\link{viewpoint_regression}}.}
}
\description{
Returns the observed costs for the different
discrete viewpoints as well as the cost of
the full viewpoint regression model.
}
