% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-viewpoint-regression.R
\name{get_discrete_weights}
\alias{get_discrete_weights}
\alias{get_discrete_weights.viewpoint_regression}
\title{Get discrete weights}
\usage{
get_discrete_weights(x)

\method{get_discrete_weights}{viewpoint_regression}(x)
}
\arguments{
\item{x}{Object from which weights should be extracted.}
}
\value{
A \code{\link[tibble]{tibble}}.
}
\description{
Extracts weights for discrete viewpoints.
}
