% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-viewpoint-regression.R
\name{get_marginal}
\alias{get_marginal}
\alias{get_marginal.viewpoint_regression}
\title{Get marginal}
\usage{
get_marginal(x, viewpoint)

\method{get_marginal}{viewpoint_regression}(x, viewpoint)
}
\arguments{
\item{x}{Object from which to extract the marginal effect.}

\item{viewpoint}{Character string identifying the viewpoint
whose marginal effect should be extracted.}
}
\value{
A \code{\link[tibble]{tibble}} with the following columns:
\itemize{
\item \code{feature_rel} - Relative feature value,
where 0 is the minimum theoretically possible feature value found when
computing the model matrix (i.e. looking over the full alphabet
of possible chord continuations),
and 1 is the maximum such value.
\item \code{feature_raw} - Absolute feature value.
\item \code{feature_obs} - \code{TRUE} if the feature value falls within
the range of actually observed feature values in the model matrix.
\item \code{effect} - The resulting effect, i.e the resulting
contribution to the model's linear predictor.
}
}
\description{
Get the marginal effect of a continuous viewpoint.
}
