% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-viewpoint-regression.R
\name{get_marginals}
\alias{get_marginals}
\alias{get_marginals.viewpoint_regression}
\title{Get marginals}
\usage{
get_marginals(x)

\method{get_marginals}{viewpoint_regression}(x)
}
\arguments{
\item{x}{Object from which to extract the marginal effects.}
}
\value{
A list with two elements:
\itemize{
\item \code{marginals} - A \code{\link[tibble]{tibble}} of marginal effects
for the continuous viewpoints, as derived by \code{\link{get_marginal}}.
\item \code{moments} - A \code{\link[tibble]{tibble}} providing
summary statistics for the different viewpoints
(mean, standard deviation, 5th percentile, 95th percentile,
minimum, maximum).
}
}
\description{
Get marginal effects of continuous viewpoints.
}
