\name{hvR-package}
\alias{hvR-package}
\alias{hvR}
\docType{package}
\title{
Hedderik van Rijn's utility functions
}
\description{
A set of utility functions, written and collected by Hedderik van Rijn. Most functions are developed in the context of analyzing experimental psychology experiments, using both behavioral and different types of psychophysical measurements.
}
\details{
\tabular{ll}{
Package: \tab hvR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-06-13\cr
License: \tab >= GPL-2\cr
}
Main purpose of this package is to have a common set of utility functions that will allow the easier exchange of code. 
}
\author{
Hedderik van Rijn

Maintainer: Hedderik van Rijn <hedderik@van-rijn.org>
}
\references{
n/a
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
