% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewpoints.R
\docType{data}
\name{hvr_viewpoints}
\alias{hvr_viewpoints}
\title{Viewpoints}
\format{
An object of class \code{list} of length 16.
}
\usage{
hvr_viewpoints
}
\description{
A default list of viewpoints for use in viewpoint regression.
}
\keyword{data}
