% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewpoints.R
\name{is_discrete}
\alias{is_discrete}
\alias{is_discrete.viewpoint}
\title{Is discrete?}
\usage{
is_discrete(x)

\method{is_discrete}{viewpoint}(x)
}
\arguments{
\item{x}{Viewpoint to check.}
}
\value{
Logical scalar.
}
\description{
Checks if a viewpoint is discrete.
}
