% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewpoints.R
\name{is_viewpoint}
\alias{is_viewpoint}
\title{Is viewpoint?}
\usage{
is_viewpoint(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
Logical scalar.
}
\description{
Checks whether an object is of class \code{viewpoint}.
}
