% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-viewpoint-regression.R
\name{is_viewpoint_regression}
\alias{is_viewpoint_regression}
\title{Testing for viewpoint regression objects}
\usage{
is_viewpoint_regression(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical scalar.
}
\description{
Tests whether a given object is a member of the class
\code{viewpoint_regression}.
}
