% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-viewpoint-regression.R
\name{list_viewpoints}
\alias{list_viewpoints}
\alias{list_viewpoints.viewpoint_regression}
\title{List viewpoints}
\usage{
list_viewpoints(x, discrete = TRUE, continuous = TRUE)

\method{list_viewpoints}{viewpoint_regression}(x, discrete = TRUE, continuous = TRUE)
}
\arguments{
\item{x}{Viewpoint regression model as created with
\code{\link{viewpoint_regression}}.}

\item{discrete}{(Logical scalar)
Whether to include discrete viewpoints.}

\item{continuous}{(Logical scalar)
Whether to include continuous viewpoints.}
}
\value{
A character vector of viewpoint identifiers.
}
\description{
Lists the viewpoints in a viewpoint regression model.
}
