% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-compute-ppm-analyses.R
\name{ltm_options}
\alias{ltm_options}
\title{LTM options}
\usage{
ltm_options(enabled = TRUE, order_bound = 10L,
  shortest_deterministic = TRUE, exclusion = TRUE,
  update_exclusion = FALSE, escape = "c")
}
\arguments{
\item{enabled}{(Logical scalar)
Whether or not the model is enabled.}

\item{order_bound}{(Integerish scalar)
The model's Markov order bound. For example, an order bound of two means
that the model makes predictions based on the two preceding symbols.}

\item{shortest_deterministic}{(Logical scalar)
If TRUE, the model will 'select' the shortest available order
that provides a deterministic prediction, if such an order exists,
otherwise defaulting to the longest available order.
For a given prediction, if this rule results in a lower model order
than would have otherwise been selected, 
then full counts (not update-excluded counts) will be used for 
the highest model order (but not for lower model orders).
This behaviour matches the implementations of PPM* in 
Pearce (2005) and
Bunton (1996).}

\item{exclusion}{(Logical scalar)
If TRUE, implements exclusion as defined in 
Pearce (2005) and
Bunton (1996).}

\item{update_exclusion}{(Logical scalar)
If TRUE, implements update exclusion as defined in 
Pearce (2005) and
Bunton (1996).}

\item{escape}{(Character scalar)
Takes values 'a', 'b', 'c', 'd', or 'ax',
corresponding to the eponymous escape methods 
in Pearce (2005).}
}
\description{
Creates an option list for the long-term discrete viewpoint
models in the viewpoint regression model.
Most (all?) of these options are passed to the PPM model.
}
