% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top-level.R
\name{model_dataset}
\alias{model_dataset}
\title{Model dataset}
\usage{
model_dataset(corpus_test, corpus_pretrain, output_dir,
  viewpoints = hvr::hvr_viewpoints, weights = NULL, poly_degree = 4L,
  max_iter = 500, corpus_test_folds = list(seq_along(corpus_test)),
  allow_repeats = FALSE, max_sample = Inf, sample_seed = 1,
  stm_opt = stm_options(), ltm_opt = ltm_options(), na_val = 0,
  perm_int = TRUE, perm_int_seed = 1, perm_int_reps = 5,
  allow_negative_weights = FALSE)
}
\arguments{
\item{corpus_test}{Corpus of chord sequences to predict,
as created by \code{\link[hrep]{corpus}}.}

\item{corpus_pretrain}{Corpus of chord sequences with which to pretrain the model,
as created by \code{\link[hrep]{corpus}}.
These chord sequences are used solely to pretrain
the discrete viewpoint models;
continuous viewpoint effects and discrete viewpoint weights
are optimised on \code{corpus_test}.}

\item{output_dir}{(Character scalar)
Directory in which to save the model outputs.}

\item{viewpoints}{List of viewpoints to apply, as created by \code{\link{new_viewpoint}}.
Defaults to a fairly comprehensive list, \code{\link{hvr_viewpoints}}.}

\item{poly_degree}{(Integer scalar)
Degree of the polynomials to compute for the continuous features.}

\item{max_iter}{(Integer scalar)
Maximum number of iterations for the optimisation routine.}

\item{corpus_test_folds}{List of cross-validation folds for applying discrete viewpoint models to
the sequences in \code{corpus_test}.
Each list element should be an integer vector indexing into \code{corpus_test}.
These integer vectors must exhaustively partition the sequences
in \code{corpus_test}.
The algorithm iterates over each fold, predicting the sequences
within that fold, and training the model using the combination of
a) the sequences from the other folds in \code{corpus_test_folds} and
b) the sequences in \code{corpus_pretrain}.
By default, there is just one fold corresponding to the
entire of \code{corpus_test}, meaning that no cross-validation is applied.}

\item{allow_repeats}{(Logical scalar)
Whether repeated chords are theoretically permitted in the
chord sequences. It is recommended to remove such repetitions
before modelling.}

\item{max_sample}{(Numeric scalar)
Maximum number of events to sample for the model matrix,
defaults to \code{Inf} (no downsampling).
Lower values of \code{max_sample} prompt random downsampling.}

\item{sample_seed}{(Integer scalar)
Random seed to make the downsampling reproducible.}

\item{stm_opt}{Options list for the short-term PPM models, as created by the function
\code{\link{stm_options}}.}

\item{ltm_opt}{Options list for the long-term PPM models, as created by the function
\code{\link{ltm_options}}.}

\item{na_val}{(Numeric scalar)
Value to use to code for NA in the model matrix.
The statistical analyses are mostly unaffected by this value.}

\item{perm_int}{(Logical scalar)
Whether to compute permutation-based feature importances.}

\item{perm_int_seed}{(Integer scalar)
Random seed for the permutation-based feature importances.}

\item{perm_int_reps}{(Integer scalar)
Number of replicates for the permutation-based feature importances
(the final estimates are averages over these replicates).}

\item{allow_negative_weights}{(Logical scalar)
Whether negative weights should be allowed for discrete features
(\code{FALSE} by default).}
}
\value{
Various model outputs are saved to \code{output_dir}.
The function returns a \code{\link[tibble]{tibble}} of predicted probabilities
for the chords in \code{corpus_test}; see
\code{\link{compute_predictions}} for an explanation of this \code{tibble}.
}
\description{
Analyses a dataset of chord sequences by constructing and optimising
a viewpoint regression model,
and using this model to generate predictions for these sequences.
}
\details{
This function wraps the following sub-routines:
\enumerate{
\item \code{\link{compute_viewpoints}}
\item \code{\link{compute_ppm_analyses}}
\item \code{\link{compute_model_matrix}}
\item \code{\link{viewpoint_regression}}
\item \code{\link{compute_predictions}}
}

Users may wish to use these sub-routines explicitly if performing
repeated analyses with different parameter settings,
to save redundant computation.
}
