% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewpoint-spec-sim.R
\name{new_spec_sim_viewpoint}
\alias{new_spec_sim_viewpoint}
\title{New spectral similarity viewpoint}
\usage{
new_spec_sim_viewpoint(half_life)
}
\arguments{
\item{half_life}{(Numeric scalar)
Half-life of the decay function, in chords.}
}
\description{
Defines a new spectral similarity viewpoint with an arbitrary half life.
}
\details{
The spectral similarity viewpoint takes the pitch-class spectral similarity
model of \insertCite{Milne2016;textual}{hrep} and adds an echoic decay
inspired by \insertCite{Leman2000;textual}{hvr}.
}
\references{
\insertAllCited{}
}
