% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewpoints-static.R
\name{new_static_viewpoint}
\alias{new_static_viewpoint}
\title{New static viewpoint}
\usage{
new_static_viewpoint(name, label, alphabet_size, discrete, mapping)
}
\arguments{
\item{name}{(Character scalar)
Name for the viewpoint (e.g. \code{pc_chord}).}

\item{label}{(Character scalar)
Label for the viewpoint (e.g. "Pitch-class chord").}

\item{alphabet_size}{For a discrete viewpoint, the number of symbols in the viewpoint's alphabet;
for a continuous viewpoint, \code{NA}.}

\item{discrete}{(Logical scalar)
Whether the viewpoint is discrete as opposed to continuous.}

\item{mapping}{An integer vector of length 24,576
(the alphabet size of \code{\link[hrep]{pc_chord}}),
where the ith element corresponds to the viewpoint value for chord i,
with chord i being encoded using the \code{\link[hrep]{pc_chord}}
encoding from the \code{hrep} package.}
}
\description{
A helper function for defining viewpoints that do not depend
on their preceding context.
}
\seealso{
\code{\link{new_viewpoint}} for defining viewpoints that do depend on
their contexts.
}
