% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewpoints.R
\name{new_viewpoint}
\alias{new_viewpoint}
\title{New viewpoint}
\usage{
new_viewpoint(name, label, alphabet_size, discrete, f_obs, f_all)
}
\arguments{
\item{name}{(Character scalar)
Name for the viewpoint (e.g. \code{pc_chord}).}

\item{label}{(Character scalar)
Label for the viewpoint (e.g. "Pitch-class chord").}

\item{alphabet_size}{For a discrete viewpoint, the number of symbols in the viewpoint's alphabet;
for a continuous viewpoint, \code{NA}.}

\item{discrete}{(Logical scalar)
Whether the viewpoint is discrete as opposed to continuous.}

\item{f_obs}{A function that takes a chord sequence as input and returns
an integer vector corresponding to the observed feature value for
each chord.
This function must accept three arguments, though
the function is free to ignore some of these arguments:
\itemize{
\item \code{chord_ids} - An integer-encoded version of the chord sequence,
using the \code{\link[hrep]{pc_chord}} encoding from the \code{hrep} package
(see \code{\link[hrep]{encode}}).
\item \code{chords} - The same chord sequence expressed as a list of
\code{\link[hrep]{pc_chord}} objects.
\item \code{verbose} - (Logical scalar) Whether or not to print
computation progress statements.
}}

\item{f_all}{A function that takes a chord sequence as input and returns
a matrix of feature values corresponding to all possible chords that
could have been observed at each point in the chord sequence.
In particular, element \code{[i, j]} should correspond to
the feature value observed if the chord sequence up to and including the
\code{i - 1}th event were to be followed by chord \code{j},
where chord \code{j} is encoded using the \code{\link[hrep]{pc_chord}}
encoding from the \code{hrep} package.
The function should accept the same argments as those listed
above for the \code{f_obs} argument.}
}
\value{
A viewpoint object that can be passed to \code{\link{compute_viewpoints}}.
}
\description{
Defines a new viewpoint.
}
\seealso{
\code{\link{new_static_viewpoint}} provides an efficient way to define
viewpoints that do not depend on the preceding chords.
}
