% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-viewpoint-regression.R
\name{plot_costs}
\alias{plot_costs}
\alias{plot_costs.viewpoint_regression}
\title{Plot costs}
\usage{
plot_costs(
  x,
  x_lab = "Cost (bits/chord)",
  y_lab = "Model",
  factor_lab = "Model type",
  factor_col = c("#E8E410", "#11A3FF", "#B50000")
)

\method{plot_costs}{viewpoint_regression}(
  x,
  x_lab = "Cost (bits/chord)",
  y_lab = "Model",
  factor_lab = "Model type",
  factor_col = c("#E8E410", "#11A3FF", "#B50000")
)
}
\arguments{
\item{x}{Viewpoint regression model as created with
\code{\link{viewpoint_regression}}.}

\item{x_lab}{Label for the x axis.}

\item{y_lab}{Label for the y axis.}

\item{factor_lab}{Label for 'model type' factor.}

\item{factor_col}{Colours for 'model type' factor.}
}
\description{
Plots observed costs for a viewpoint regression
as extracted with \code{\link{get_costs}}.
}
