% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-viewpoint-regression.R
\name{plot_discrete_weights}
\alias{plot_discrete_weights}
\alias{plot_discrete_weights.viewpoint_regression}
\title{Plot weights for discrete viewpoints}
\usage{
plot_discrete_weights(x, x_lab = "Viewpoint", y_lab = "Weight",
  colours = c("#B50000", "#11A3FF"), labels = x$viewpoint_labels)

\method{plot_discrete_weights}{viewpoint_regression}(x,
  x_lab = "Viewpoint", y_lab = "Weight", colours = c("#B50000",
  "#11A3FF"), labels = x$viewpoint_labels)
}
\arguments{
\item{x}{Viewpoint regression model object,
as created by \code{\link{viewpoint_regression}}.}

\item{x_lab}{Label for x axis.}

\item{y_lab}{Label for y axis.}

\item{colours}{Vector of colours to use.}

\item{labels}{\code{\link[tibble]{tibble}} providing
textual labels for the viewpoints.}
}
\description{
Plots regression weights for discrete viewpoints
in a viewpoint regression model.
}
