% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-viewpoint-regression.R
\name{plot_marginals}
\alias{plot_marginals}
\title{Plot marginals}
\usage{
plot_marginals(model_1, model_2 = NULL, model_labels = NULL,
  x_lab = "Feature value", y_lab = "Effect",
  viewpoint_labels = x$viewpoint_labels, fill = "blue", alpha = 0.25,
  scales = "free", ...)
}
\arguments{
\item{model_1}{Viewpoint regression model,
as created by \code{\link{viewpoint_regression}}.}

\item{model_2}{Optional second viewpoint regression model,
to be plotted for comparison.}

\item{model_labels}{Only relevant if a second viewpoint regression model is provided,
in which case this should be a character vector of length 2
corresponding to the plot labels for these two models.}

\item{x_lab}{(Character scalar)
Label for the x axis.}

\item{y_lab}{(Character scalar)
Label for the y axis.}

\item{viewpoint_labels}{\code{\link[tibble]{tibble}} of viewpoint labels.}

\item{fill}{(Character scalar)
Fill colour for the shaded region identifying the 5th-95th percentiles
of observed feature values.}

\item{alpha}{(Numeric scalar)
Alpha value for the shaded region.}

\item{scales}{Passed to \code{\link[ggplot2]{facet_wrap}}.}

\item{...}{Further arguments to pass to \code{\link[ggplot2]{facet_wrap}}.}
}
\description{
Plots the marginal effects for continuous viewpoints in a
viewpoint regression model, as extracted by
\code{\link{get_marginal}}.
}
\details{
The x axis spans the 5th-95th percentiles
of theoretically possible feature values
as computed in the derivation of the model matrix
(see \code{\link{compute_model_matrix}}).
The shaded area identifies the 5th-95th percentiles
of observed feature values when computing the model matrix.
}
