% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-viewpoint-regression.R
\name{viewpoint_regression}
\alias{viewpoint_regression}
\title{Conduct viewpoint regression}
\usage{
viewpoint_regression(
  parent_dir,
  max_iter = 500,
  perm_int = TRUE,
  perm_int_seed = 1,
  perm_int_reps = 5,
  allow_negative_weights = FALSE,
  viewpoint_dir = file.path(parent_dir, "0-viewpoints"),
  model_matrix_dir = file.path(parent_dir, "2-model-matrix"),
  output_dir = file.path(parent_dir, "3-viewpoint-regression")
)
}
\arguments{
\item{parent_dir}{(Character scalar)
The parent directory for the output files, shared with functions such as
\code{\link{compute_viewpoints}} and \code{\link{compute_ppm_analyses}}.
Ignored if all other directory arguments are manually specified.}

\item{max_iter}{(Integer scalar)
Maximum number of iterations for the optimisation routine.}

\item{perm_int}{(Logical scalar)
Whether to compute permutation-based feature importances.}

\item{perm_int_seed}{(Integer scalar)
Random seed for the permutation-based feature importances.}

\item{perm_int_reps}{(Integer scalar)
Number of replicates for the permutation-based feature importances
(the final estimates are averages over these replicates).}

\item{allow_negative_weights}{(Logical scalar)
Whether negative weights should be allowed for discrete features
(\code{FALSE} by default).}

\item{viewpoint_dir}{(Character scalar)
The directory for the already-generated
output files from \code{\link{compute_viewpoints}}.
The default should be correct if the user used the
default \code{dir} argument in \code{\link{compute_viewpoints}}.}

\item{model_matrix_dir}{(Character scalar)
The directory for the already-generated
output files from \code{\link{compute_model_matrix}}.
The default should be correct if the user used the
default \code{dir} argument in \code{\link{compute_model_matrix}}.}

\item{output_dir}{(Character scalar)
The output directory for the new results.
Will be created if it doesn't exist already.}
}
\value{
The primary output is a viewpoint regression model object
written to disk in the \code{dir} directory.
This object contains the fitted viewpoint regression weights
and feature importance analyses.
Various plots may be constructed from this object:
\itemize{
\item \code{\link{plot_costs}}
\item \code{\link{plot_perm_int}}
\item \code{\link{plot_marginals}}
\item \code{\link{plot_discrete_weights}}
}
}
\description{
Fits a viewpoint regression model.
The following routines must be run first:
\enumerate{
\item \code{\link{compute_viewpoints}}
\item \code{\link{compute_ppm_analyses}}
\item \code{\link{compute_model_matrix}}
}
}
\details{
The optimisation method is "BFGS" if \code{allow_negative_weights}
is \code{TRUE} and "L-BFGS-B" otherwise (see \code{\link[stats]{optim}}).
}
