% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configurations.R
\docType{data}
\name{vr_config}
\alias{vr_config}
\title{Viewpoint regression configurations}
\format{
An object of class \code{list} of length 1.
}
\usage{
vr_config
}
\description{
Provides a list of viewpoint regression model configurations
as optimised in previous work.
}
\details{
\itemize{
\item \code{hvr_1} - Model optimised to predict the sixth chord in short
chord sequences extracted from the Billboard popular music dataset
(\code{\link[hcorp]{popular_1}}).
}
}
\keyword{data}
