% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-compute-ppm-analyses.R
\name{compute_ppm_analyses}
\alias{compute_ppm_analyses}
\title{Compute PPM analyses}
\usage{
compute_ppm_analyses(parent_dir, viewpoint_dir = file.path(parent_dir,
  "0-viewpoints"), output_dir = file.path(parent_dir, "1-ppm"),
  stm_opt = stm_options(), ltm_opt = ltm_options(),
  seq_test_folds = list(readRDS(file.path(viewpoint_dir,
  "about.rds"))$seq_test),
  seq_pretrain = readRDS(file.path(viewpoint_dir,
  "about.rds"))$seq_pretrain,
  viewpoints = readRDS(file.path(viewpoint_dir,
  "about.rds"))$discrete_viewpoints)
}
\description{
This function computes PPM analyses.
}
