% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hvr-package.R
\docType{package}
\name{hvr-package}
\alias{hvr}
\alias{hvr-package}
\title{hvr: Harmonic Viewpoint Regression}
\description{
Tools for constructing Harmonic Viewpoint Regression (HVR) models.
}
\author{
\strong{Maintainer}: Peter Harrison \email{pmc.harrison@gmail.com}

}
\keyword{internal}
