% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-viewpoint-regression.R
\name{plot_perm_int}
\alias{plot_perm_int}
\alias{plot_perm_int.viewpoint_regression}
\title{Plot permutation-based feature importances}
\usage{
plot_perm_int(
  x,
  labels = x$viewpoint_labels,
  axis_label = "Feature importance (bits/chord)",
  order_by_label = FALSE,
  error_bars = FALSE,
  discrete_label = "Discrete",
  continuous_label = "Continuous"
)

\method{plot_perm_int}{viewpoint_regression}(
  x,
  labels = x$viewpoint_labels,
  axis_label = "Feature importance (bits/chord)",
  order_by_label = FALSE,
  error_bars = FALSE,
  discrete_label = "Discrete",
  continuous_label = "Continuous"
)
}
\arguments{
\item{x}{Viewpoint regression model as created with
\code{\link{viewpoint_regression}}.}

\item{labels}{\code{\link[tibble]{tibble}} of viewpoint labels.}

\item{axis_label}{(Character scalar)
Axis label.}

\item{order_by_label}{(Logical scalar)
If \code{TRUE}, bars are ordered by viewpoint label,
otherwise they are ordered by quantity.}

\item{error_bars}{(Logical scalar)
Whether or not to plot error bars, which will correspond to the
2.5th/97.5th percentiles of the permutation-importance replicates.
These will only be stable for relatively large numbers of replicates
(50+).}

\item{discrete_label}{Label to apply to discrete viewpoints.}

\item{continuous_label}{Label to apply to continuous viewpoints.}
}
\description{
Plots permutation-based feature importances
for a viewpoint regression model as extracted
from the \code{$perm_int} slot.
}
